refactor: refactor benefit module strcuture
