+++ modules/benefit/dtos/request/benefitDtoRequest.js
@@ -1,0 +1,9 @@
+class BenefitDTORequest {
+  constructor(name, imageUrl, imagePublicId) {
+    this.name = name;
+    this.imageUrl = imageUrl;
+    this.imagePublicId = imagePublicId;
+  }
+}
+
+module.exports = BenefitDTORequest;
