+++ modules/benefit/dtos/response/benefitDtoResponse.js
@@ -1,0 +1,12 @@
+class BenefitDTOResponse {
+  constructor(data) {
+    this.id = data.id;
+    this.name = data.name;
+    this.imageUrl = data.imageUrl;
+    this.createdAt = data.createdAt;
+    this.updatedAt = data.updatedAt;
+    this.deletedAt = data.deletedAt;
+  }
+}
+
+module.exports = BenefitDTOResponse;
